feat: refactoring the about page
