+++ app/about/page.tsx
@@ -1,309 +1,18 @@
 "use client"
-import { useState, useEffect, JSX } from "react"
-import { createClientComponentClient } from "@supabase/auth-helpers-nextjs"
-import Image from "next/image"
-import Link from "next/link"
-import { motion } from "framer-motion"
-import { ChevronDown, ChevronUp, ExternalLink } from "lucide-react"
-import * as LucideIcons from "lucide-react"
 import { Header } from "@/components/header"
 import Footer from "@/components/footer"
 import { Button } from "@/components/ui/button"
-import { TeamMembersTab } from "@/components/admin/about/TeamMembersTab"
+import { useAboutData } from "@/hooks/useAboutData"
+import { defaultAboutData } from "@/constants/about/about"
+import { HeroSection } from "@/components/about/heroSection"
+import { MissionSection } from "@/components/about/missionSection"
+import { ValuesSection } from "@/components/about/valuesSection"
+import { TeamSection } from "@/components/about/teamSection"
+import { FAQSection } from "@/components/about/faqSection"
 
-// Types
-type AboutPage = {
-  hero_title: string
-  hero_subtitle: string
-  mission_title: string
-  mission_content: string
-  values_title: string
-  values_subtitle: string
-  team_title: string
-  team_subtitle: string
-  faq_title: string
-  faq_subtitle: string
-}
-
-type CoreValue = {
-  id: string
-  title: string
-  description: string
-  icon_name: string
-  color: string
-  order_position: number
-  is_active: boolean
-}
-
-type TeamMember = {
-  id: string
-  name: string
-  role: string
-  bio: string
-  image_url: string | null
-  twitter_url: string | null
-  linkedin_url: string | null
-  instagram_url: string | null
-  order_position: number
-  is_active: boolean
-}
-
-type FAQ = {
-  id: string
-  question: string
-  answer: string
-  order_position: number
-  is_active: boolean
-}
-
-export default function AboutPage() { 
-  // Estados para los datos
-  const [aboutData, setAboutData] = useState<AboutPage | null>(null)
-  const [coreValues, setCoreValues] = useState<CoreValue[]>([])
-  const [teamMembers, setTeamMembers] = useState<TeamMember[]>([])
-  const [faqs, setFaqs] = useState<FAQ[]>([])
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState("")
-
-  // Estado para controlar los acordeones de FAQ
-  const [openFaq, setOpenFaq] = useState<number | null>(0)
-
-  const supabase = createClientComponentClient()
-
-  // Fetch todos los datos al cargar la página
-  useEffect(() => {
-    fetchAllData()
-  }, [])
-
-  const getIconComponent = (iconName: string) => {
-    // Convertir el nombre a PascalCase (ej: "heart" -> "Heart", "user-plus" -> "UserPlus")
-    const iconComponentName = iconName
-      .split('-')
-      .map(word => word.charAt(0).toUpperCase() + word.slice(1))
-      .join('')
-
-    // Obtener el componente del ícono
-    const IconComponent = (LucideIcons as any)[iconComponentName]
-
-    // Si no existe, usar Shield como default
-    return IconComponent || LucideIcons.Shield
-  }
-
-  const fetchAllData = async () => {
-    setIsLoading(true)
-    try {
-      // Fetch about page general info
-      const { data: aboutPageData, error: aboutError } = await supabase
-        .from('about_page')
-        .select('*')
-        .single()
-
-      if (aboutError && aboutError.code !== 'PGRST116') {
-        console.error('About page error:', aboutError)
-      } else {
-        setAboutData(aboutPageData)
-      }
-
-      // Fetch core values (solo activos)
-      const { data: valuesData, error: valuesError } = await supabase
-        .from('core_values')
-        .select('*')
-        .eq('is_active', true)
-        .order('order_position')
-
-      if (valuesError) {
-        console.error('Values error:', valuesError)
-      } else {
-        setCoreValues(valuesData || [])
-      }
-
-      // Fetch team members (solo activos)
-      const { data: teamData, error: teamError } = await supabase
-        .from('team_members')
-        .select('*')
-        .eq('is_active', true)
-        .order('order_position')
-
-      if (teamError) {
-        console.error('Team error:', teamError)
-      } else {
-        setTeamMembers(teamData || [])
-      }
-
-      // Fetch FAQs (solo activos)
-      const { data: faqData, error: faqError } = await supabase
-        .from('faqs')
-        .select('*')
-        .eq('is_active', true)
-        .order('order_position')
-
-      if (faqError) {
-        console.error('FAQ error:', faqError)
-      } else {
-        setFaqs(faqData || [])
-      }
-
-    } catch (error: any) {
-      console.error('Error fetching data:', error)
-      setError('Failed to load page data')
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  // Función para obtener ícono SVG por nombre
-  const getIconSvg = (iconName: string) => {
-    const icons: { [key: string]: JSX.Element } = {
-      shield: (
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M9 12l2 2 4-4m5.618-4.016A11.955 11.955 0 0112 2.944a11.955 11.955 0 01-8.618 3.04A12.02 12.02 0 003 9c0 5.591 3.824 10.29 9 11.622 5.176-1.332 9-6.03 9-11.622 0-1.042-.133-2.052-.382-3.016z"
-        />
-      ),
-      lightbulb: (
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M9.663 17h4.673M12 3v1m6.364 1.636l-.707.707M21 12h-1M4 12H3m3.343-5.657l-.707-.707m2.828 9.9a5 5 0 117.072 0l-.548.547A3.374 3.374 0 0014 18.469V19a2 2 0 11-4 0v-.531c0-.895-.356-1.754-.988-2.386l-.548-.547z"
-        />
-      ),
-      users: (
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M17 20h5v-2a3 3 0 00-5.356-1.857M17 20H7m10 0v-2c0-.656-.126-1.283-.356-1.857M7 20H2v-2a3 3 0 015.356-1.857M7 20v-2c0-.656.126-1.283.356-1.857m0 0a5.002 5.002 0 019.288 0M15 7a3 3 0 11-6 0 3 3 0 016 0zm6 3a2 2 0 11-4 0 2 2 0 014 0zM7 10a2 2 0 11-4 0 2 2 0 014 0z"
-        />
-      ),
-      target: (
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8s3.59-8 8-8 8 3.59 8 8-3.59 8-8 8zm0-14c-3.31 0-6 2.69-6 6s2.69 6 6 6 6-2.69 6-6-2.69-6-6-6zm0 10c-2.21 0-4-1.79-4-4s1.79-4 4-4 4 1.79 4 4-1.79 4-4 4z"
-        />
-      ),
-      heart: (
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M20.84 4.61a5.5 5.5 0 0 0-7.78 0L12 5.67l-1.06-1.06a5.5 5.5 0 0 0-7.78 7.78l1.06 1.06L12 21.23l7.78-7.78 1.06-1.06a5.5 5.5 0 0 0 0-7.78z"
-        />
-      )
-    }
-
-    return icons[iconName] || icons.shield
-  }
-
-  // Valores por defecto si no hay datos en la base de datos
-  const defaultAboutData: AboutPage = {
-    hero_title: "About Game Tested Tech",
-    hero_subtitle: "Your trusted source for honest gaming hardware reviews, guides, and tech insights since 2025.",
-    mission_title: "Our Mission",
-    mission_content: "At Game Tested Tech, we believe that gamers deserve honest, thorough, and unbiased reviews of the technology that powers their passion. Our mission is to cut through marketing hype and provide real-world insights based on extensive testing in actual gaming scenarios.",
-    values_title: "Our Core Values",
-    values_subtitle: "These principles guide everything we do at Game Tested Tech",
-    team_title: "Meet Our Team",
-    team_subtitle: "The passionate gamers and tech experts behind Game Tested Tech",
-    faq_title: "Frequently Asked Questions",
-    faq_subtitle: "Find answers to common questions about our review process, team, and mission."
-  }
-
-  // Usar datos por defecto si no hay datos de la base de datos
+export default function AboutPage() {
+  const { aboutData, coreValues, teamMembers, faqs, isLoading, error } = useAboutData()
   const pageData = aboutData || defaultAboutData
-
-  // Variantes de animación
-  const containerVariants = {
-    hidden: { opacity: 0 },
-    visible: {
-      opacity: 1,
-      transition: {
-        staggerChildren: 0.1,
-        delayChildren: 0.3,
-      },
-    },
-  }
-
-  const itemVariants = {
-    hidden: { y: 20, opacity: 0 },
-    visible: {
-      y: 0,
-      opacity: 1,
-      transition: { type: "spring", stiffness: 100 },
-    },
-  }
-
-  // Función para guardar/actualizar team member (NO maneja imágenes, eso lo hace el componente)
-  const saveTeamMember = async (member: Partial<TeamMember>) => {
-    try {
-      if (member.id) {
-        // Update existing
-        const { error } = await supabase
-          .from('team_members')
-          .update({
-            ...member,
-            updated_at: new Date().toISOString()
-          })
-          .eq('id', member.id)
-
-        if (error) throw error
-      } else {
-        // Create new
-        const { error } = await supabase
-          .from('team_members')
-          .insert([{
-            ...member,
-            order_position: teamMembers.length
-          }])
-
-        if (error) throw error
-      }
-
-      fetchAllData()
-      setSuccess("Team member saved successfully!")
-
-    } catch (error: any) {
-      setError("Failed to save team member: " + error.message)
-    }
-  }
-
-  // Función para eliminar elementos
-  const deleteItem = async (table: string, id: string) => {
-    try {
-      const { error } = await supabase
-        .from(table)
-        .delete()
-        .eq('id', id)
-
-      if (error) throw error
-
-      fetchAllData()
-      setSuccess("Item deleted successfully!")
-
-    } catch (error: any) {
-      setError("Failed to delete item: " + error.message)
-    }
-  }
-
-  // Función para toggle active/inactive
-  const toggleActive = async (table: string, id: string, currentStatus: boolean) => {
-    try {
-      const { error } = await supabase
-        .from(table)
-        .update({ is_active: !currentStatus })
-        .eq('id', id)
-
-      if (error) throw error
-      fetchAllData()
-
-    } catch (error: any) {
-      setError("Failed to update status: " + error.message)
-    }
-  }
 
   if (error) {
     return (
@@ -311,12 +20,7 @@ export default function AboutPage() {
         <div className="text-center">
           <h1 className="text-2xl font-bold text-white mb-4">Error Loading Page</h1>
           <p className="text-gray-400 mb-6">{error}</p>
-          <Button
-            onClick={() => window.location.reload()}
-            className="bg-[#9d8462] hover:bg-[#8d7452] text-white"
-          >
-            Try Again
-          </Button>
+          <Button onClick={() => window.location.reload()}>Try Again</Button>
         </div>
       </div>
     )
@@ -325,317 +29,40 @@ export default function AboutPage() {
   return (
     <div className="min-h-screen bg-[#0f0f23] text-gray-200">
       <Header />
-
-      {/* Hero Section */}
-      <div className="relative">
-        <div className="absolute inset-0">
-          <div className="absolute inset-0 bg-gradient-to-b from-[#1a1a2e]/80 to-[#0a0a14]"></div>
-        </div>
-
-        <div className="relative pt-20 pb-24 px-4 sm:px-6 lg:px-8 max-w-7xl mx-auto">
-          <motion.div
-            className="text-center"
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.6 }}
-          >
-            <h1 className="text-4xl md:text-5xl font-bold text-white mb-6">
-              {pageData.hero_title}
-            </h1>
-            <p className="text-xl text-gray-300 max-w-3xl mx-auto">
-              {pageData.hero_subtitle}
-            </p>
-          </motion.div>
-        </div>
-        <div className="absolute bottom-0 left-0 w-full h-16 bg-gradient-to-t from-[#0a0a14] to-transparent"></div>
-      </div>
-
-      {/* Our Mission Section */}
-      <section className="py-16 bg-[#0f0f23]">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="grid grid-cols-1 lg:grid-cols-2 gap-12 items-center">
-            <motion.div
-              initial={{ opacity: 0, x: -30 }}
-              whileInView={{ opacity: 1, x: 0 }}
-              transition={{ duration: 0.6 }}
-              viewport={{ once: true }}
-            >
-              <h2 className="text-3xl font-bold text-white mb-6">{pageData.mission_title}</h2>
-              <div className="space-y-4 text-gray-300">
-                {pageData.mission_content.split('\n').map((paragraph, index) => (
-                  <p key={index}>{paragraph}</p>
-                ))}
-              </div>
-            </motion.div>
-
-            <motion.div
-              className="relative"
-              initial={{ opacity: 0, x: 30 }}
-              whileInView={{ opacity: 1, x: 0 }}
-              transition={{ duration: 0.6, delay: 0.2 }}
-              viewport={{ once: true }}
-            >
-              <div className="aspect-video relative rounded-lg overflow-hidden shadow-xl">
-                <Image
-                  src="/images/game-controller-logo.png"
-                  alt="Gaming setup being tested"
-                  fill
-                  className="object-contain max-w-[500px] max-h-[600px] m-auto"
-                />
-              </div>
-              <div className="absolute -bottom-6 -right-6 w-48 h-48 bg-[#9d8462]/20 rounded-full blur-2xl"></div>
-              <div className="absolute -top-6 -left-6 w-48 h-48 bg-[#0673b8]/20 rounded-full blur-2xl"></div>
-            </motion.div>
-          </div>
-        </div>
-      </section>
-
-      {/* Our Values Section */}
+      
+      <HeroSection title={pageData.hero_title} subtitle={pageData.hero_subtitle} />
+      
+      <MissionSection 
+        title={pageData.mission_title} 
+        content={pageData.mission_content} 
+      />
+      
       {coreValues.length > 0 && (
-        <section className="py-16 bg-[#0a0a14]">
-          <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <motion.div
-              className="text-center mb-16"
-              initial={{ opacity: 0, y: 20 }}
-              whileInView={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.6 }}
-              viewport={{ once: true }}
-            >
-              <h2 className="text-3xl font-bold text-white mb-4">{pageData.values_title}</h2>
-              <p className="text-xl text-gray-300 max-w-3xl mx-auto">
-                {pageData.values_subtitle}
-              </p>
-            </motion.div>
-
-            {/* Flex container que se mantiene centrado automáticamente */}
-            <div className="flex flex-wrap justify-center gap-8 max-w-6xl mx-auto">
-              {coreValues.map((value, index) => (
-                <motion.div
-                  key={value.id}
-                  className="bg-[#1a1a2e] rounded-lg p-6 shadow-lg w-full sm:w-80 md:w-80 lg:w-80 flex-shrink-0"
-                  initial={{ opacity: 0, y: 20 }}
-                  whileInView={{ opacity: 1, y: 0 }}
-                  transition={{ duration: 0.6, delay: index * 0.1 }}
-                  viewport={{ once: true }}
-                  whileHover={{ y: -5 }}
-                >
-                  <div
-                    className="w-12 h-12 rounded-full flex items-center justify-center mb-4"
-                    style={{ backgroundColor: `${value.color}20` }}
-                  >
-                    {(() => {
-                      const IconComponent = getIconComponent(value.icon_name)
-                      return (
-                        <IconComponent
-                          className="h-6 w-6"
-                          style={{ color: value.color }}
-                        />
-                      )
-                    })()}
-                  </div>
-                  <h3 className="text-xl font-bold text-white mb-2">{value.title}</h3>
-                  <p className="text-gray-300">{value.description}</p>
-                </motion.div>
-              ))}
-            </div>
-          </div>
-        </section>
+        <ValuesSection 
+          title={pageData.values_title}
+          subtitle={pageData.values_subtitle}
+          values={coreValues}
+        />
       )}
-
-      {/* Meet Our Team Section */}
+      
       {teamMembers.length > 0 && (
-        <section className="py-16 bg-[#0f0f23]">
-          <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <motion.div
-              className="text-center mb-16"
-              initial={{ opacity: 0, y: 20 }}
-              whileInView={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.6 }}
-              viewport={{ once: true }}
-            >
-              <h2 className="text-3xl font-bold text-white mb-4">{pageData.team_title}</h2>
-              <p className="text-xl text-gray-300 max-w-3xl mx-auto">
-                {pageData.team_subtitle}
-              </p>
-            </motion.div>
-
-            {/* Flex container que se mantiene centrado automáticamente */}
-            <div className="flex flex-wrap justify-center gap-8 max-w-6xl mx-auto">
-              {teamMembers.map((member, index) => (
-                <motion.div
-                  key={member.id}
-                  className="bg-[#1a1a2e] rounded-lg overflow-hidden shadow-lg w-full sm:w-80 md:w-80 lg:w-80 flex-shrink-0"
-                  initial={{ opacity: 0, y: 20 }}
-                  whileInView={{ opacity: 1, y: 0 }}
-                  transition={{ duration: 0.6, delay: index * 0.1 }}
-                  viewport={{ once: true }}
-                  whileHover={{ y: -5 }}
-                >
-                  {member.image_url && (
-                    <div className="relative aspect-square">
-                      <Image
-                        src={member.image_url}
-                        alt={member.name}
-                        fill
-                        className="object-cover object-top"
-                      />
-                    </div>
-                  )}
-                  <div className="p-6">
-                    <h3 className="text-xl font-bold text-white mb-1">{member.name}</h3>
-                    <p className="text-[#9d8462] mb-3">{member.role}</p>
-                    {member.bio && (
-                      <p className="text-gray-300 text-sm mb-4">{member.bio}</p>
-                    )}
-                    <div className="flex gap-3">
-                      {member.twitter_url && (
-                        <a
-                          href={member.twitter_url}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="text-gray-400 hover:text-white transition-colors"
-                          aria-label={`${member.name}'s Twitter`}
-                        >
-                          <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            className="h-5 w-5"
-                            fill="currentColor"
-                            viewBox="0 0 24 24"
-                          >
-                            <path d="M24 4.557c-.883.392-1.832.656-2.828.775 1.017-.609 1.798-1.574 2.165-2.724-.951.564-2.005.974-3.127 1.195-.897-.957-2.178-1.555-3.594-1.555-3.179 0-5.515 2.966-4.797 6.045-4.091-.205-7.719-2.165-10.148-5.144-1.29 2.213-.669 5.108 1.523 6.574-.806-.026-1.566-.247-2.229-.616-.054 2.281 1.581 4.415 3.949 4.89-.693.188-1.452.232-2.224.084.626 1.956 2.444 3.379 4.6 3.419-2.07 1.623-4.678 2.348-7.29 2.04 2.179 1.397 4.768 2.212 7.548 2.212 9.142 0 14.307-7.721 13.995-14.646.962-.695 1.797-1.562 2.457-2.549z" />
-                          </svg>
-                        </a>
-                      )}
-                      {member.linkedin_url && (
-                        <a
-                          href={member.linkedin_url}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="text-gray-400 hover:text-white transition-colors"
-                          aria-label={`${member.name}'s LinkedIn`}
-                        >
-                          <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            className="h-5 w-5"
-                            fill="currentColor"
-                            viewBox="0 0 24 24"
-                          >
-                            <path d="M4.98 3.5c0 1.381-1.11 2.5-2.48 2.5s-2.48-1.119-2.48-2.5c0-1.38 1.11-2.5 2.48-2.5s2.48 1.12 2.48 2.5zm.02 4.5h-5v16h5v-16zm7.982 0h-4.968v16h4.969v-8.399c0-4.67 6.029-5.052 6.029 0v8.399h4.988v-10.131c0-7.88-8.922-7.593-11.018-3.714v-2.155z" />
-                          </svg>
-                        </a>
-                      )}
-                      {member.instagram_url && (
-                        <a
-                          href={member.instagram_url}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="text-gray-400 hover:text-white transition-colors"
-                          aria-label={`${member.name}'s Instagram`}
-                        >
-                          <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            className="h-5 w-5"
-                            fill="currentColor"
-                            viewBox="0 0 24 24"
-                          >
-                            <path d="M12 2.163c3.204 0 3.584.012 4.85.07 3.252.148 4.771 1.691 4.919 4.919.058 1.265.069 1.645.069 4.849 0 3.205-.012 3.584-.069 4.849-.149 3.225-1.664 4.771-4.919 4.919-1.266.058-1.644.07-4.85.07-3.204 0-3.584-.012-4.849-.07-3.26-.149-4.771-1.699-4.919-4.92-.058-1.265-.07-1.644-.07-4.849 0-3.204.013-3.583.07-4.849.149-3.227 1.664-4.771 4.919-4.919 1.266-.057 1.645-.069 4.849-.069zm0-2.163c-3.259 0-3.667.014-4.947.072-4.358.2-6.78 2.618-6.98 6.98-.059 1.281-.073 1.689-.073 4.948 0 3.259.014 3.668.072 4.948.2 4.358 2.618 6.78 6.98 6.98 1.281.058 1.689.072 4.948.072 3.259 0 3.668-.014 4.948-.072 4.354-.2 6.782-2.618 6.979-6.98.059-1.28.073-1.689.073-4.948 0-3.259-.014-3.667-.072-4.947-.196-4.354-2.617-6.78-6.979-6.98-1.281-.059-1.69-.073-4.949-.073zm0 5.838c-3.403 0-6.162 2.759-6.162 6.162s2.759 6.163 6.162 6.163 6.162-2.759 6.162-6.163c0-3.403-2.759-6.162-6.162-6.162zm0 10.162c-2.209 0-4-1.79-4-4 0-2.209 1.791-4 4-4s4 1.791 4 4c0 2.21-1.791 4-4 4zm6.406-11.845c-.796 0-1.441.645-1.441 1.44s.645 1.44 1.441 1.44c.795 0 1.439-.645 1.439-1.44s-.644-1.44-1.439-1.44z" />
-                          </svg>
-                        </a>
-                      )}
-                    </div>
-                  </div>
-                </motion.div>
-              ))}
-            </div>
-          </div>
-        </section>
+        <TeamSection 
+          title={pageData.team_title}
+          subtitle={pageData.team_subtitle}
+          members={teamMembers}
+        />
       )}
-
-      {/* FAQ Section */}
+      
       {faqs.length > 0 && (
-        <section className="py-16 bg-[#0a0a14]">
-          <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <div className="grid grid-cols-1 lg:grid-cols-3 gap-12">
-              <motion.div
-                className="lg:col-span-1"
-                initial={{ opacity: 0, x: -20 }}
-                whileInView={{ opacity: 1, x: 0 }}
-                transition={{ duration: 0.6 }}
-                viewport={{ once: true }}
-              >
-                <h2 className="text-3xl font-bold text-white mb-6">{pageData.faq_title}</h2>
-                <p className="text-gray-300 mb-8">
-                  {pageData.faq_subtitle}
-                </p>
-                <div className="hidden lg:block">
-                  <Link
-                    href="/contact"
-                    className="inline-flex items-center gap-2 bg-[#ff6b35] hover:bg-[#ff8c5a] text-white px-6 py-3 rounded-md transition-all transform hover:scale-105"
-                  >
-                    Have more questions? <ExternalLink className="h-4 w-4" />
-                  </Link>
-                </div>
-              </motion.div>
-
-              <motion.div
-                className="lg:col-span-2 space-y-4"
-                variants={containerVariants}
-                initial="hidden"
-                whileInView="visible"
-                viewport={{ once: true }}
-              >
-                {faqs.map((faq, index) => (
-                  <motion.div
-                    key={faq.id}
-                    className="bg-[#1a1a2e] rounded-lg overflow-hidden"
-                    variants={itemVariants}
-                  >
-                    <button
-                      className="w-full flex justify-between items-center p-6 text-left focus:outline-none"
-                      onClick={() => setOpenFaq(openFaq === index ? null : index)}
-                    >
-                      <h3 className="text-lg font-medium text-white">{faq.question}</h3>
-                      {openFaq === index ? (
-                        <ChevronUp className="h-5 w-5 text-gray-400" />
-                      ) : (
-                        <ChevronDown className="h-5 w-5 text-gray-400" />
-                      )}
-                    </button>
-                    <div
-                      className={`px-6 pb-6 transition-all duration-300 ease-in-out ${openFaq === index ? "block opacity-100" : "hidden opacity-0"
-                        }`}
-                    >
-                      <p className="text-gray-300">{faq.answer}</p>
-                    </div>
-                  </motion.div>
-                ))}
-              </motion.div>
-
-              <div className="lg:hidden col-span-1 text-center">
-                <Link
-                  href="/contact"
-                  className="inline-flex items-center gap-2 bg-[#ff6b35] hover:bg-[#ff8c5a] text-white px-6 py-3 rounded-md transition-all duration-300 border-0 transform hover:scale-105"
-                >
-                  Have more questions? <ExternalLink className="h-4 w-4" />
-                </Link>
-              </div>
-            </div>
-          </div>
-        </section>
+        <FAQSection 
+          title={pageData.faq_title}
+          subtitle={pageData.faq_subtitle}
+          faqs={faqs}
+        />
       )}
-
-      {/* Loading State */}
-      {isLoading && (
-        <section className="py-16 bg-[#0f0f23]">
-          <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <div className="flex items-center justify-center min-h-[400px]">
-              <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-[#9d8462]"></div>
-            </div>
-          </div>
-        </section>
-      )}
-
-      {/* Footer */}
+      
+      {isLoading && <div>Loading...</div>}
+      
       <Footer />
     </div>
   )
