+++ constants/about/about.ts
@@ -1,0 +1,14 @@
+import { AboutPage } from '@/types/about/about'
+
+export const defaultAboutData: AboutPage = {
+  hero_title: "About Game Tested Tech",
+  hero_subtitle: "Your trusted source for honest gaming hardware reviews, guides, and tech insights since 2025.",
+  mission_title: "Our Mission",
+  mission_content: "At Game Tested Tech, we believe that gamers deserve honest, thorough, and unbiased reviews of the technology that powers their passion. Our mission is to cut through marketing hype and provide real-world insights based on extensive testing in actual gaming scenarios.",
+  values_title: "Our Core Values",
+  values_subtitle: "These principles guide everything we do at Game Tested Tech",
+  team_title: "Meet Our Team",
+  team_subtitle: "The passionate gamers and tech experts behind Game Tested Tech",
+  faq_title: "Frequently Asked Questions",
+  faq_subtitle: "Find answers to common questions about our review process, team, and mission."
+}
