+++ types/home/type.ts
@@ -1,4 +1,3 @@
-// types/home/index.ts
 export interface Author {
   name: string;
   avatar_url: string;
